[API] Revert 85d66af, introduce IActivityProducer and -Consumer instead
Thanks to Stefan Rossbach for pointing out a less intrusive way than
renaming the IActivityProvider interface to
IActivityProviderAndConsumer.

This patch deprecates the IActivityProvider interface and, in order to
take a first step towards the separation of concerns of Activity
producing and consuming, adds two new interfaces each of which is now
implemented by the AbstractActivityProvider class.

I left a few TODOs in the code that are to be addressed by Arsenij
Solovjev.

Change-Id: I5beaa2086aaeeedd5306687220d0ab87492c95da