[INTERNAL] Copy large portions of Saros/E to Saros/I
This is only temporary. These classes are largely equal to their
Saros/Eclipse counterparts. However, there are some parts that are
commented out, others have been slightly changed to avoid compilation
errors. Please note that this code is not intended for execution, but
will serve as a "copy-paste" source.

The extension classes were copied to the package "net.internal", and
do not reside under the "core" package. The reason is simple:
The package name is used to unmarshall the XMPP packets, so the
package names need to be identical for all participants, regardless of
their IDE.

Further proceeding:
- Classes from the "core" package will be moved (fully or partially) to
  the "intellij" package (not to be confused with the "core" or
  "intellij" *project*).
- Once a class made it into the "intellij" package, it can be
  consolidated with its Saros/Eclipse counterpart, which should reduce
  code duplication and let the Saros/Core grow.
- Eventually, the "core" package inside Saros/IntelliJ should be empty/
  deleted.

Change-Id: Ia6debff47d5b24d774a83c6744687ceaa16099ad